Add mockito endpoint testing.

+++ inventory-backend/src/test/java/com/revature/inventory/service/impl/ProductServiceImplTest.java
@@ -12,6 +12,8 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
+import org.junit.Before;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.runner.RunWith;
@@ -160,7 +162,9 @@ class ProductServiceImplTest {
 
         verify(productDao, atLeastOnce()).deleteById(longCaptor.getValue());
 
-        assertEquals(10L, longCaptor.getValue().intValue());
+        //assertEquals(10L, longCaptor.getValue().intValue());
+        assertEquals(10L, longCaptor.getValue().longValue());
+
 	}
 
 }
